utils: move LoggedErrorStack helper to worker/provisioner
The only usage of utils.LoggedErrorStack was the provisioner worker.
Move this type and its tests to worker/provisioner/logger.go
