+++ worker/provisioner/logging.go
@@ -1,23 +1,20 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package utils
+package provisioner
 
 import (
 	"github.com/juju/errors"
-	"github.com/juju/loggo"
 	"github.com/juju/utils/featureflag"
 
 	"github.com/juju/juju/feature"
 )
 
-var logger = loggo.GetLogger("juju.utils")
-
-// LoggedErrorStack is a developer helper function that will cause the error
+// loggedErrorStack is a developer helper function that will cause the error
 // stack of the error to be printed out at error severity if and only if the
 // "log-error-stack" feature flag has been specified.  The passed in error
 // is also the return value of this function.
-func LoggedErrorStack(err error) error {
+func loggedErrorStack(err error) error {
 	if featureflag.Enabled(feature.LogErrorStack) {
 		logger.Errorf("error stack:\n%s", errors.ErrorStack(err))
 	}
