+++ worker/provisioner/logging_test.go
@@ -1,7 +1,7 @@
 // Copyright 2015 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package utils_test
+package provisioner
 
 import (
 	"errors"
@@ -11,7 +11,6 @@ import (
 
 	"github.com/juju/juju/feature"
 	"github.com/juju/juju/testing"
-	"github.com/juju/juju/utils"
 )
 
 type logSuite struct {
@@ -22,7 +21,7 @@ var _ = gc.Suite(&logSuite{})
 
 func (*logSuite) TestFlagNotSet(c *gc.C) {
 	err := errors.New("test error")
-	err2 := utils.LoggedErrorStack(err)
+	err2 := loggedErrorStack(err)
 	c.Assert(err, gc.Equals, err2)
 	c.Assert(c.GetTestLog(), gc.Equals, "")
 }
@@ -30,8 +29,8 @@ func (*logSuite) TestFlagNotSet(c *gc.C) {
 func (s *logSuite) TestFlagSet(c *gc.C) {
 	s.SetFeatureFlags(feature.LogErrorStack)
 	err := errors.New("test error")
-	err2 := utils.LoggedErrorStack(err)
+	err2 := loggedErrorStack(err)
 	c.Assert(err, gc.Equals, err2)
-	expected := "ERROR juju.utils error stack:\ntest error"
+	expected := "ERROR juju.provisioner error stack:\ntest error"
 	c.Assert(c.GetTestLog(), jc.Contains, expected)
 }
